perf: convert TOML config to JSON
Since a separate JS module is needed to parse the TOML file, by using the JSON format we can avoid that additional download.

The size of both the TOML parser and of the config file is negligible, but one less dependency sounds nice and JSON is the native format for the web anyway.